+++ test/index.test.js
@@ -11,9 +11,23 @@ const os = require('os');
 let initInstance;
 let testingObject;
 let addedObject = {
-    "add": {
-        "a": "a",
-        "b": "b"
+    "happyCase": {
+        "aws_access_key_id": "a",
+        "aws_secret_access_key": "b"
+    },
+    "emptyObject": {
+    },
+    "missingOneKey": {
+        "aws_access_key_id": "a",
+    },
+    "containsOneExtraKey": {
+        "aws_access_key_id": "a",
+        "aws_secret_access_key": "b",
+        "extra": 123
+    },
+    "wrongKeys": {
+        "wrong": "a",
+        "keys": "b"
     }
 };
 let childObject = {
@@ -24,6 +38,8 @@ let childObject = {
 };
 
 describe('V2 awsProfileHandler unit test', () => {
+    const TEST_STRING = 'test';
+
     beforeEach(() => {
         testingObject = {
             "default": childObject.default,
@@ -32,7 +48,8 @@ describe('V2 awsProfileHandler unit test', () => {
             }
         };
 
-        Utils.readFile.mockReturnValue(null);
+
+        Utils.readFile.mockReturnValue(TEST_STRING);
         Ini.decodeIniData.mockReturnValue(testingObject);
         Utils.deepCopy.mockImplementation((object) => {
             return JSON.parse(JSON.stringify(object))
@@ -40,19 +57,28 @@ describe('V2 awsProfileHandler unit test', () => {
     });
 
     describe('listProfile', () => {
-        test('with default credentials path', () => {
+        it('with default credentials path', () => {
             const defaultFilePath = path.join(os.homedir(), '.aws', 'credentials');
             let result = awsProfileHandler.listProfiles();
             expect(Utils.readFile).toBeCalledWith(defaultFilePath);
-            expect(Ini.decodeIniData).toBeCalled();
+            expect(Ini.decodeIniData).toBeCalledWith(TEST_STRING);
             expect(result).toEqual(['default', 'awesome']);
         });
-        test('with customized credentials path', () => {
+        it('with customized credentials path', () => {
             const customFilePath = "custom/file/path";
             let result = awsProfileHandler.listProfiles(customFilePath);
             expect(Utils.readFile).toBeCalledWith(customFilePath);
-            expect(Ini.decodeIniData).toBeCalled();
+            expect(Ini.decodeIniData).toBeCalledWith(TEST_STRING);
             expect(result).toEqual(['default', 'awesome']);
+        });
+
+        it('when file is empty return empty array', () => {
+            const customFilePath = "custom/file/path";
+            Ini.decodeIniData.mockReturnValue({});
+            let result = awsProfileHandler.listProfiles(customFilePath);
+            expect(Utils.readFile).toBeCalledWith(customFilePath);
+            expect(Ini.decodeIniData).toBeCalledWith(TEST_STRING);
+            expect(result).toEqual([]);
         });
     });
 
@@ -66,27 +92,140 @@ describe('V2 awsProfileHandler unit test', () => {
             let result = awsProfileHandler.getProfileCredentials("default");
             expect(result).toEqual(childObject.default);
         });
+
+        it('should throw error if input profile is missing', () => {
+            function invalidInput() {
+                return awsProfileHandler.getProfileCredentials();
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: profile name cannot be omitted nor only contains white spaces.');
+        });
+
+        it('should throw error if input profile is only white space', () => {
+            function invalidInput() {
+                return awsProfileHandler.getProfileCredentials("\n\t  ");
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: profile name cannot be omitted nor only contains white spaces.');
+        });
+
     });
 
-    test('deleteProfile', () => {
-        awsProfileHandler.deleteProfile("awesome");
-        expect(Ini.encodeIniFormat)
-            .toBeCalledWith(childObject);
-        expect(Utils.writeFile).toBeCalled();
+    describe('deleteProfile', () => {
+        it('happy case', () => {
+            awsProfileHandler.deleteProfile("awesome");
+            expect(Ini.encodeIniFormat)
+                .toBeCalledWith(childObject);
+            expect(Utils.writeFile).toBeCalled();
+        });
+
+        it('delete the last object', () => {
+            testingObject = {
+                "awesome": {
+                    "aaron": "that's me"
+                }
+            };
+            awsProfileHandler.deleteProfile("awesome");
+            expect(Ini.encodeIniFormat)
+                .toBeCalledWith(childObject);
+            expect(Utils.writeFile).toBeCalled();
+        });
+
+        it('the input profile doesn\'t exist', () => {
+            awsProfileHandler.deleteProfile("amazing");
+            expect(Ini.encodeIniFormat)
+                .toBeCalledWith(childObject);
+            expect(Utils.writeFile).toBeCalled();
+        });
+
+        it('should throw error if input profile is missing', () => {
+            function invalidInput() {
+                return awsProfileHandler.deleteProfile();
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: profile name cannot be omitted nor only contains white spaces.');
+        });
+
+        it('should throw error if input profile is only white space', () => {
+            function invalidInput() {
+                return awsProfileHandler.deleteProfile("\n\t  ");
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: profile name cannot be omitted nor only contains white spaces.');
+        });
     });
 
-    test('addProfile', () => {
-        let expected_result = {
-            "default": childObject.default,
-            "awesome": {
-                "aaron": "that's me"
-            },
-            "add": addedObject.add
-        };
-        awsProfileHandler.addProfile("add", addedObject.add);
-        expect(Utils.writeFile).toBeCalled();
-        expect(Ini.encodeIniFormat)
-            .toBeCalledWith(expected_result);
+    describe('addProfile', () => {
+        it ('happy case', () => {
+            let expected_result = {
+                "default": childObject.default,
+                "awesome": {
+                    "aaron": "that's me"
+                },
+                "add": addedObject.happyCase
+            };
+            awsProfileHandler.addProfile("add", addedObject.happyCase);
+            expect(Utils.writeFile).toBeCalled();
+            expect(Ini.encodeIniFormat)
+                .toBeCalledWith(expected_result);
+        });
+
+
+        it('should throw error if input profile is missing', () => {
+            function invalidInput() {
+                return awsProfileHandler.addProfile();
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: profile name cannot be omitted nor only contains white spaces.');
+        });
+
+        it('should throw error if input profile is only white space', () => {
+            function invalidInput() {
+                return awsProfileHandler.addProfile("\n\t  ");
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: profile name cannot be omitted nor only contains white spaces.');
+        });
+
+        it('should throw error if input credentials is missing', () => {
+            function invalidInput() {
+                return awsProfileHandler.addProfile('profile');
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: credentials cannot be omitted nor empty.');
+        });
+
+        it('should throw error if input credentials is empty', () => {
+            function invalidInput() {
+                return awsProfileHandler.addProfile('profile', addedObject.emptyObject);
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid Input: credentials cannot be omitted nor empty.');
+        });
+
+        it('should throw error if input credentials is missing one key', () => {
+            function invalidInput() {
+                return awsProfileHandler.addProfile('profile', addedObject.missingOneKey);
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid input: credentials schema is invalid.');
+        });
+
+        it('should throw error if input credentials has extra keys', () => {
+            function invalidInput() {
+                return awsProfileHandler.addProfile('profile', addedObject.containsOneExtraKey);
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid input: credentials schema is invalid.');
+        });
+
+        it('should throw error if input credentials object has wrong keys', () => {
+            function invalidInput() {
+                return awsProfileHandler.addProfile('profile', addedObject.wrongKeys);
+            }
+            expect(invalidInput)
+                .toThrowError('Invalid input: credentials schema is invalid.');
+        });
     })
 });