+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.0.0] - 2018-02-16
+### Added
+- Validator: Input profile name cannot be white space, '\n', '\t', etc...
+- Validator: Input profile cannot be missing.
+- Validator: Credential schema has to be valid.
+
+### Changed
+- If the file is empty, the package returns an empty array instead of throwing an error.
+
 
 ## [2.0.0] - 2018-02-16
 ### Added
@@ -18,3 +27,5 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - All instance methods: listProfiles, getProfile, deleteProfile, and addProfile.
 
 [2.0.0]: https://github.com/novking/aws-profile-handler/compare/v1.1.0...v2.0.0
+[2.0.1]: https://github.com/novking/aws-profile-handler/compare/v2.0.0...v2.0.1
+